+++ packages/gamut/src/Tip/InfoTip/InfoTipButton.tsx
@@ -22,15 +22,26 @@ export type InfoTipButtonProps = ComponentProps<typeof InfoTipButtonBase> &
   Pick<InfoTipProps, 'emphasis'>;
 
 export const InfoTipButton = forwardRef<ButtonBaseElements, InfoTipButtonProps>(
-  ({ active, children, emphasis, ...props }, ref) => {
+  (
+    {
+      active,
+      children,
+      emphasis,
+      'aria-label': ariaLabel,
+      'aria-labelledby': ariaLabelledby,
+      ...props
+    },
+    ref
+  ) => {
     const Icon = emphasis === 'high' ? MiniInfoCircleIcon : MiniInfoOutlineIcon;
 
     return (
       <InfoTipButtonBase
+        {...props}
         active={active}
         aria-expanded={active}
-        aria-label="Show information"
-        {...props}
+        aria-label={ariaLabel}
+        aria-labelledby={ariaLabelledby}
         ref={ref}
       >
         {Icon && (